extended `shell.h` file with the required functionalities:
Explanation of the changes:

Added Function Prototypes:

I added the function prototypes for the built-in commands `shell_cd`, `shell_exit`, and `shell_env`. These functions will handle the respective built-in commands.

Removed Some Function Prototypes:

I noticed that some of the function prototypes were already defined in the original snippet. As they are already present in the `shell.h file`, I did not include them again.

Changed Function Definitions to Prototypes:

I removed the function definitions from `shell.h`, leaving only the function prototypes. The actual function definitions should be implemented in the source file (`shell_main.c`) to maintain separation of concerns.

Updated the Comment:

I updated the comment at the top of the file to reflect the additional functionalities and built-in commands we added.